# Small optimization to cctz year_index.